Revert "Refactor ManageVaultView to take components as params."
This reverts commit ab3612b13ae40d8546cbfb21072f5becd0173d45.
