+++ features/borrow/manage/containers/ManageVaultView.tsx
@@ -1,34 +1,113 @@
+import { Icon } from '@makerdao/dai-ui-icons'
 import { trackingEvents } from 'analytics/analytics'
 import { useAppContext } from 'components/AppContextProvider'
-import { DefaultVaultHeaderProps } from 'components/vault/DefaultVaultHeader'
+import { TextWithCheckmark } from 'components/TextWithCheckmark'
+import { DefaultVaultHeader } from 'components/vault/DefaultVaultHeader'
+import { VaultAllowanceStatus } from 'components/vault/VaultAllowance'
+import { VaultChangesWithADelayCard } from 'components/vault/VaultChangesWithADelayCard'
+import { VaultFormContainer } from 'components/vault/VaultFormContainer'
+import { VaultProxyStatusCard } from 'components/vault/VaultProxy'
+import { ManageVaultFormHeader } from 'features/borrow/manage/containers/ManageVaultFormHeader'
 import { VaultHistoryEvent } from 'features/vaultHistory/vaultHistory'
+import { VaultHistoryView } from 'features/vaultHistory/VaultHistoryView'
 import { useTranslation } from 'next-i18next'
 import React, { useEffect } from 'react'
-import { Box, Grid } from 'theme-ui'
+import { Box, Divider, Grid, Text } from 'theme-ui'
 
+import { ManageVaultCollateralAllowance } from '../../../../components/vault/commonMultiply/ManageVaultCollateralAllowance'
+import { ManageVaultDaiAllowance } from '../../../../components/vault/commonMultiply/ManageVaultDaiAllowance'
+import { VaultErrors } from '../../../../components/vault/VaultErrors'
+import { VaultWarnings } from '../../../../components/vault/VaultWarnings'
 import { ManageVaultState } from '../pipes/manageVault'
 import { createManageVaultAnalytics$ } from '../pipes/manageVaultAnalytics'
+import { ManageVaultButton } from './ManageVaultButton'
+import { ManageVaultConfirmation, ManageVaultConfirmationStatus } from './ManageVaultConfirmation'
+import { ManageVaultDetails } from './ManageVaultDetails'
+import { ManageVaultEditing } from './ManageVaultEditing'
 
-export interface ManageVaultContainerProps {
-  manageVault: ManageVaultState
-  vaultHistory: VaultHistoryEvent[]
+function ManageVaultMultiplyTransition({ stage, vault }: ManageVaultState) {
+  const { t } = useTranslation()
+  return stage === 'multiplyTransitionEditing' ? (
+    <Grid mt={-3}>
+      <Grid variant="text.paragraph3" sx={{ color: 'text.subtitle' }}>
+        <TextWithCheckmark>
+          {t('borrow-to-multiply.checkmark1', { token: vault.token.toUpperCase() })}
+        </TextWithCheckmark>
+        <TextWithCheckmark>{t('borrow-to-multiply.checkmark2')}</TextWithCheckmark>
+        <TextWithCheckmark>{t('borrow-to-multiply.checkmark3')}</TextWithCheckmark>
+        <TextWithCheckmark>{t('borrow-to-multiply.checkmark4')}</TextWithCheckmark>
+      </Grid>
+      <Divider />
+      <Grid gap={2}>
+        <Text variant="paragraph2" sx={{ fontWeight: 'semiBold' }}>
+          {t('borrow-to-multiply.subheader2')}
+        </Text>
+        <Text variant="paragraph3" sx={{ color: 'text.subtitle' }}>
+          {t('borrow-to-multiply.paragraph2')}
+        </Text>
+      </Grid>
+    </Grid>
+  ) : (
+    <Box>
+      <Icon name="multiply_transition" size="auto" />
+    </Box>
+  )
 }
 
-interface ManageVaultContainerComponents {
-  header: (props: DefaultVaultHeaderProps) => JSX.Element
-  details: (props: ManageVaultState) => JSX.Element
-  form: (props: ManageVaultState) => JSX.Element
-  history: (props: Pick<ManageVaultContainerProps, 'vaultHistory'>) => JSX.Element
+function ManageVaultForm(props: ManageVaultState & { isInstiVault?: boolean }) {
+  const {
+    isEditingStage,
+    isProxyStage,
+    isCollateralAllowanceStage,
+    isDaiAllowanceStage,
+    isManageStage,
+    isMultiplyTransitionStage,
+    accountIsConnected,
+    daiAllowanceTxHash,
+    collateralAllowanceTxHash,
+    vault: { token },
+    stage,
+  } = props
+
+  return (
+    <VaultFormContainer toggleTitle="Edit Vault">
+      <ManageVaultFormHeader {...props} />
+      {isEditingStage && <ManageVaultEditing {...props} />}
+      {isCollateralAllowanceStage && <ManageVaultCollateralAllowance {...props} />}
+      {isDaiAllowanceStage && <ManageVaultDaiAllowance {...props} />}
+      {isManageStage && <ManageVaultConfirmation {...props} />}
+      {isMultiplyTransitionStage && <ManageVaultMultiplyTransition {...props} />}
+      {accountIsConnected && (
+        <>
+          <VaultErrors {...props} />
+          <VaultWarnings {...props} />
+          {stage === 'manageSuccess' && <VaultChangesWithADelayCard />}
+          <ManageVaultButton {...props} />
+        </>
+      )}
+      {isProxyStage && <VaultProxyStatusCard {...props} />}
+      {isCollateralAllowanceStage && (
+        <VaultAllowanceStatus
+          {...props}
+          allowanceTxHash={collateralAllowanceTxHash}
+          token={token}
+        />
+      )}
+      {isDaiAllowanceStage && (
+        <VaultAllowanceStatus {...props} allowanceTxHash={daiAllowanceTxHash} token={'DAI'} />
+      )}
+      {isManageStage && <ManageVaultConfirmationStatus {...props} />}
+    </VaultFormContainer>
+  )
 }
 
 export function ManageVaultContainer({
   manageVault,
   vaultHistory,
-  header: Header,
-  details: Details,
-  form: Form,
-  history: History,
-}: ManageVaultContainerProps & ManageVaultContainerComponents) {
+}: {
+  manageVault: ManageVaultState
+  vaultHistory: VaultHistoryEvent[]
+}) {
   const { manageVault$, context$ } = useAppContext()
   const {
     vault: { id, ilk },
@@ -52,14 +131,14 @@ export function ManageVaultContainer({
 
   return (
     <>
-      <Header header={t('vault.header', { ilk, id })} id={id} ilkData={ilkData} />
+      <DefaultVaultHeader header={t('vault.header', { ilk, id })} id={id} ilkData={ilkData} />
       <Grid variant="vaultContainer">
         <Grid gap={5} mb={[0, 5]}>
-          <Details {...manageVault} />
-          <History vaultHistory={vaultHistory} />
+          <ManageVaultDetails {...manageVault} />
+          <VaultHistoryView vaultHistory={vaultHistory} />
         </Grid>
         <Box>
-          <Form {...manageVault} />
+          <ManageVaultForm {...manageVault} />
         </Box>
       </Grid>
     </>
